requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
# K-Prism Interactive Segmentation — Python dependencies
# Python 3.10  |  CUDA 11.7  |  Tested on Linux (Ubuntu 20.04 / 22.04)
#
# Install with:
#   pip install -r requirements.txt
#
# NOTE: PyTorch and detectron2 must be installed SEPARATELY — see README.md.
# ─────────────────────────────────────────────────────────────────────────────

# Deep-learning core (installed separately — see README)
# torch==1.13.1
# torchvision==0.14.1

# Web UI
streamlit==1.10.0
streamlit-drawable-canvas==0.9.3
# Streamlit 1.10 imports altair.vegalite.v4 — removed in Altair 5+
altair>=4.2.0,<5

# Medical image I/O
SimpleITK==2.3.1
monai==0.9.0

# Image processing
Pillow==10.3.0
opencv-python==4.9.0.80
scipy==1.13.0
albumentations==1.4.6

# Numerical
numpy==1.26.4

# Config / utilities
omegaconf==2.3.0
hydra-core==1.3.2
einops==0.8.0
PyYAML==6.0.1
tqdm==4.66.4
tensorboardX==2.6.2.2

# detectron2 / fvcore installed separately — see README
# detectron2==0.6
# fvcore==0.1.5.post20221221
